@@ -1,27 +1,33 @@
-import type { ComponentEvents, ComponentProps, SvelteComponent } from 'svelte'
+import type { Component, ComponentEvents, ComponentProps } from 'svelte'
 import type { Readable } from 'svelte/store'
 
-export type RenderConfig<TComponent extends SvelteComponent = SvelteComponent> =
+export type RenderConfig<TComponent extends Component = Component> =
     | ComponentRenderConfig<TComponent>
     | string
     | number
     | Readable<string | number>
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export type Constructor<TInstance> = new (...args: any[]) => TInstance
-
-export class ComponentRenderConfig<TComponent extends SvelteComponent = SvelteComponent> {
+export class ComponentRenderConfig<TComponent extends Component = Component> {
     constructor(
-        public component: Constructor<TComponent>,
-        public props?: ComponentProps<TComponent> | Readable<ComponentProps<TComponent>>
+        public component: Component,
+        public props?: Record<string, unknown>
     ) {}
 
+    /**
+     * @deprecated This method will be removed in the next major release. Please use svelte5 event syntax instead.
+     */
     eventHandlers: [keyof ComponentEvents<TComponent>, (ev: Event) => void][] = []
+
+    /**
+     * @deprecated This method will be removed in the next major release. Please use svelte5 event syntax instead.
+     */
     on<TEventType extends keyof ComponentEvents<TComponent>>(
         type: TEventType,
         handler: (ev: ComponentEvents<TComponent>[TEventType]) => void
     ): this {
         this.eventHandlers.push([type, handler])
+        this.props ??= {} as ComponentProps<TComponent>
+        ;(this.props as Record<string, unknown>)[`on${String(type)}`] = handler
         return this
     }
 
@@ -33,18 +39,18 @@ export class ComponentRenderConfig<TComponent extends SvelteComponent = SvelteCo
 }
 
 // Allow omission of the `props` argument if the component accepts no props.
-export function createRender<TComponent extends SvelteComponent<Record<string, never>>>(
-    component: Constructor<TComponent>
+export function createRender<TComponent extends Component<any>>(
+    component: TComponent
 ): ComponentRenderConfig<TComponent>
 
-export function createRender<TComponent extends SvelteComponent>(
-    component: Constructor<TComponent>,
+export function createRender<TComponent extends Component<any>>(
+    component: TComponent,
     props: ComponentProps<TComponent> | Readable<ComponentProps<TComponent>>
 ): ComponentRenderConfig<TComponent>
 
-export function createRender<TComponent extends SvelteComponent>(
-    component: Constructor<TComponent>,
+export function createRender<TComponent extends Component<any>>(
+    component: TComponent,
     props?: ComponentProps<TComponent> | Readable<ComponentProps<TComponent>>
 ): ComponentRenderConfig<TComponent> {
-    return new ComponentRenderConfig(component, props)
+    return new ComponentRenderConfig(component, props as Record<string, unknown>)
 }
