@@ -1,2 +1,2 @@
-export { default as Render } from './Render.svelte'
 export * from './createRender.js'
+export { default as Render } from './Render.svelte'
