@@ -1,7 +1,7 @@
 <script lang="ts">
     import { Render, type RenderConfig } from '$lib/index.js'
-
-    type TComponent = $$Generic<SvelteComponent>
+    import type { Component } from 'svelte'
+    type TComponent = $$Generic<Component>
 
     export let config: RenderConfig<TComponent>
 </script>
