@@ -1,7 +1,7 @@
 import { sveltekit } from '@sveltejs/kit/vite'
 import { defineConfig } from 'vitest/config'
 
-export default defineConfig({
+export default defineConfig(({ mode }) => ({
     plugins: [sveltekit()],
     test: {
         include: ['src/**/*.test.ts'],
@@ -13,5 +13,8 @@ export default defineConfig({
             exclude: ['docs/**', '.trunk/**', '.svelte-kit/**', 'tests/**', 'src/routes/**']
         },
         reporters: ['verbose', ['junit', { outputFile: './junit-vitest.xml' }]]
+    },
+    resolve: {
+        conditions: mode === 'test' ? ['browser'] : []
     }
-})
+}))
